requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

CairoSVG==2.7.0
flet==0.8.2
numpy==1.25.1
opencv-python==4.8.0.74
Pillow==10.0.0
platformdirs==3.9.1
